+++ core/lib/preload.d.ts
@@ -2,7 +2,7 @@
 import { WriteStream, ReadStream } from 'fs-extra';
 import { Pack, Extract } from 'tar-fs';
 export declare type HPROC = {
-    getInfo: () => HINFO;
+    getInfo: () => Promise<HINFO>;
     existsSync: (path: string) => Promise<boolean>;
     copySync: (path_from: string, path_to: string) => void;
     removeSync: (path: string) => Promise<void>;