+++ core/lib/sn/Button.d.ts
@@ -16,7 +16,7 @@ export declare class Button extends Container {
     private sp_b_pic;
     private sp_pic;
     constructor(hArg: HArg, evtMng: IEvtMng, resolve: () => void, canFocus: () => boolean);
-    drawDesignCast(gdc: IGenerateDesignCast): void;
+    drawDesignCast(_gdc: IGenerateDesignCast): void;
     cvsResize(): void;
     update_b_pic(fn: string, txt: Text): void;
     private loaded_b_pic;